backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.12
python-escpos==3.1
reportlab==4.2.5
requests==2.32.3
schedule==1.2.2
pystray==0.19.5
Pillow==10.4.0
pywin32==308
aiofiles==24.1.0
httpx==0.27.2
